% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{get_a_roc}
\alias{get_a_roc}
\title{get_a_roc}
\source{
Obuchowski NA. Fundamentals of clinical research for radiologists:
  ROC Analysis. American Journal of Roentgenology. 2005;184.
}
\usage{
get_a_roc(AUC)
}
\arguments{
\item{AUC}{Area under the ROC curve. Ranges from greater than 0.5 to 1.}
}
\description{
Function to calculate the A binormal distribution parameter
}
\examples{

  get_a_roc(AUC = 0.6)

}
