% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deff.R
\name{get_deff}
\alias{get_deff}
\title{get_deff}
\usage{
get_deff(x, cluster, size)
}
\arguments{
\item{x}{variable to calculate \code{ICC} from}

\item{cluster}{variable identifying the clusters or groupings of the variable}

\item{size}{number of samples within a cluster}
}
\value{
A numeric value for design effect (DEFF)
}
\description{
Function to calculate design effect (\code{DEFF}) given intracluster
correlation coefficient (\code{ICC}) and cluster size.
}
\examples{
  #

}
