% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepped_wedge.r
\name{get_deff_swedge}
\alias{get_deff_swedge}
\title{Function to calculate the design effect of a stepped wedge cluster randomised
study.}
\usage{
get_deff_swedge(k, b, t, icc, size)
}
\arguments{
\item{k}{Number of steps in the stepped wedge study}

\item{b}{Number of baseline measurements}

\item{t}{Number of measurements after each step}

\item{icc}{Intracluster correlation coefficient (\code{ICC})}

\item{size}{cluster size}
}
\value{
A numeric value of the design effect of a stepped wedge cluster
    ranomised study
}
\description{
Function to calculate the design effect of a stepped wedge cluster randomised
study.
}
\examples{

  get_deff_swedge(k = 2, b = 1, t = 1, icc = 0.034, size = 192)

}
