% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_icc.R
\name{get_icc}
\alias{get_icc}
\title{get_icc}
\usage{
get_icc(x, cluster)
}
\arguments{
\item{x}{variable to calculate ICC from}

\item{cluster}{variable identifying the clusters or groupings of the variable}
}
\value{
A vector with named elements n (total number of non-missing observations),
    \code{clusters} (number of clusters after deleting missing data),
    \code{rho} (intra-cluster correlation), and \code{deff} (design effect).
}
\description{
Function to calculate the intracluster correlation coefficient of an
indicator collected from random cluster survey (RCS). This is a wrapper of
the \code{deff()} function in the \code{Hmisc} package.
}
\examples{
  x <- sample(1:2, size = 25, replace = TRUE)
  cluster <- c(rep(1, 5), rep(2, 5), rep(3, 5), rep(4, 5), rep(5, 5))
  get_icc(x = x, cluster = cluster)

}
