% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{get_k_roc}
\alias{get_k_roc}
\title{get_k_roc}
\source{
Obuchowski NA. Fundamentals of clinical research for radiologists:
  ROC Analysis. American Journal of Roentgenology. 2005;184.
}
\usage{
get_k_roc(p)
}
\arguments{
\item{p}{Prevalence or proportion of cases in the relevant population}
}
\description{
Function to calculate ratio of cases to non-cases in the sample for an ROC
analysis
}
\examples{

  get_k_roc(p = 0.5)

}
