% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{get_nc_roc}
\alias{get_nc_roc}
\title{Function to calculate the number of cases needed for an ROC analysis for a
single test}
\source{
Obuchowski NA. Fundamentals of clinical research for radiologists:
  ROC Analysis. American Journal of Roentgenology. 2005;184.
}
\usage{
get_nc_roc(z = 1.96, precision = 0.05, vf)
}
\arguments{
\item{z}{The z-score/z-value for a 95\% confidence interval. Typical vaues
for \code{z} are 1.645 for a 90\% confidence interval, 1.75 for a 92\%
confidence interval, 1.96 for a 95\% confidence interval and 2.05 for a
96\% confidence interval. Default is 1.96.}

\item{precision}{The numeric value of the desired precision. Default is set
at ±5\% (0.05).}

\item{vf}{Variance function of the area under the curve (AUC) for the
received operator characteristic (ROC) surve sample size calculations}
}
\value{
A numeric value for the number of cases needed for an ROC analysis
    for a single test
}
\description{
Function to calculate the number of cases needed for an ROC analysis for a
single test
}
\examples{

  get_nc_roc(z = 1.96, precision = 0.05, vf = 0.00991024)

}
