% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepped_wedge.r
\name{get_ss_swedge}
\alias{get_ss_swedge}
\title{get_ss_swedge}
\usage{
get_ss_swedge(z1 = 1.96, z2 = 0.84, p1, p2, deff)
}
\arguments{
\item{z1}{The z-score indicating level of significance (confidence interval).
Default at 1.96 for 95\% confidence interval.}

\item{z2}{desired power. Default at 0.84 for 80\% power.}

\item{p1}{Expected proportion for control group.}

\item{p2}{Expected proportion for intervention group}

\item{deff}{Design effect of a stepped wedge cluster randomised study}
}
\value{
A numeric value of sample size needed for a stepped wedge cluster
    ranomised study
}
\description{
Function to calculate sample size requirements for a stepped wedge cluster
randomised study.
}
\examples{

 get_ss_swedge(z1 = 1.96, z2 = 0.84,
               p1 = 0.2, p2 = 0.15,
               deff = get_deff_swedge(k = 2, b = 1, t = 1,
                                      icc = 0.034, size = 192))

}
