% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{get_vf_roc}
\alias{get_vf_roc}
\title{Function to calculate variance function (VF) of the area under the curve
(AUC) for the receiver operator characteristic (ROC) curve sample size
calculations}
\source{
Obuchowski NA. Fundamentals of clinical research for radiologists:
  ROC Analysis. American Journal of Roentgenology. 2005;184.
}
\usage{
get_vf_roc(A, p)
}
\arguments{
\item{A}{Binormal distribution parameter}

\item{p}{Prevalence or proportion of cases in the relevant population. Used
to calculate ratio \code{k} of cases to non-cases in the relevant
population}
}
\value{
A numeric value for variance function (\code{VF}) of the area under
    the curve (AUC) for the receiver operator characteristic (ROC) curve
    sample size calculations
}
\description{
Function to calculate variance function (VF) of the area under the curve
(AUC) for the receiver operator characteristic (ROC) curve sample size
calculations
}
\examples{

  get_vf_roc(A = 2, p = 0.5)

}
