% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampsizer.R
\docType{package}
\name{sampsizer}
\alias{sampsizer}
\alias{sampsizer-package}
\title{sampsizer}
\description{
Calculate sample sizes for various survey settings and conditions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ernestguevarra/sampsizer}
  \item Report bugs at \url{https://github.com/ernestguevarra/sampsizer/issues}
}

}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} [copyright holder]

}
\keyword{internal}
