% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deff.R
\name{get_deff}
\alias{get_deff}
\title{get_deff}
\usage{
get_deff(x, cluster, size)
}
\arguments{
\item{x}{variable to calculate \code{ICC} from}

\item{cluster}{variable identifying the clusters or groupings of the variable}

\item{size}{number of samples within a cluster}
}
\value{
A numeric value for design effect (DEFF)
}
\description{
Function to calculate design effect (\code{DEFF}) given intracluster
correlation coefficient (\code{ICC}) and cluster size.
}
\examples{
  x <- sample(1:2, size = 25, replace = TRUE)
  cluster <- c(rep(1, 5), rep(2, 5), rep(3, 5), rep(4, 5), rep(5, 5))
  get_deff(x = x, cluster = cluster, size = 10)

}
