% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence.R
\name{get_ss_prevalence}
\alias{get_ss_prevalence}
\title{get_ss_prevalence}
\usage{
get_ss_prevalence(z = 1.96, p = 0.5, c = 0.1, deff = 2)
}
\arguments{
\item{z}{The z-score/z-value for a 95\% confidence interval. Typical vaues
for \code{z} are 1.645 for a 90\% confidence interval, 1.75 for a 92\%
confidence interval, 1.96 for a 95\% confidence interval and 2.05 for a
96\% confidence interval. Default is 1.96.}

\item{p}{The numeric value of the expected prevalence/proportion. Default is
0.5 (for 50\% prevalence/proportion) as this gives the highest sample
size estimate. If previous prevalence/proportion is known, this value
can be used for \code{p}.}

\item{c}{The numeric value of the desired precision of the resulting
prevalence/proportion estimate. For coverage surveys, default is set at
±10\% (0.1).}

\item{deff}{The numeric for design effect (\code{deff}) to account for
suvey design. For simple random sample (SRS) surveys, a \code{deff} of
1 is used. For a random cluster survey (RCS), a \code{deff} of 2 is
usually used. If survey data from previous cluster surveys is available,
\code{deff} can be calculated using \code{deff()} function from
\code{Hmisc} package.}
}
\value{
A numeric value for sample size required to estimate the required
    prevalence/proportion.
}
\description{
Function to calculate sample size for estimating prevalence/proportional
indicators from a simple random sample (SRS) or random cluster sample (RCS)
surveys.
}
\examples{
  # Calculate sample size to estimate a proportion with 95\% confidence
  # interval, an expected proporition of 50\%, precision of 10\% and a design
  # effect of 2
  get_ss_prevalence(z = 1.96, p = 0.5, c = 0.1, deff = 2)

}
